+++ charts/sessionspaces/templates/clusterrole-visit-member.yaml
@@ -1,0 +1,20 @@
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: visit-member
+rules:
+- apiGroups: ["argoproj.io"]
+  resources: [
+    "eventsources",
+    "sensors",
+    "workflows",
+    "workfloweventbindings",
+    "workflowtemplates",
+    "clusterworkflowtemplates",
+    "cronworkflows",
+    "workflowtaskresults",
+    ]
+  verbs: ["get", "watch", "list"] 
+- apiGroups: ["argoproj.io"]
+  resources: ["workflows"]
+  verbs: ["create"]
