+++ sessionspaces/src/permissionables/mod.rs
@@ -1,6 +1,8 @@
+/// gidNumber for sessions
+mod ldap;
 /// Beamline sessions
 mod session;
 /// Associations between subjects and sessions
 mod subject_session;
 
-pub use self::{session::Session, subject_session::SubjectSession};
+pub use self::{ldap::ldap_search, session::Session, subject_session::SubjectSession};
