+++ sessionspaces/src/resources/mod.rs
@@ -1,15 +1,9 @@
 /// The config map for kyverno policy
 mod config_maps;
-/// The visit-member ServiceAccount and corresponding Role
-mod member_sa;
 /// The Namespace for a beamline session
 mod namespace;
-/// The argo-workflows ServiceAccount and corresponding Role
-mod workflows_sa;
 
 pub use self::{
     config_maps::create_configmap,
-    member_sa::create_visit_member_role,
     namespace::{create_namespace, delete_namespace},
-    workflows_sa::create_argo_workflows_role,
 };
